requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
xgboost
joblib
pydantic
matplotlib
seaborn
python-dotenv
python-multipart
